CollisionDetection and Movement
Added:
-Player movement
-PlayerCar movement functions
-PlayerCar collision functions
-Collision via keyboard movement.

Modified:
-Spawner list is now public
-Rect is now a polygon

+++ 3DExample1/src/GameObject.cpp
@@ -14,10 +14,10 @@ GameObject::GameObject(): _posX(0), _posY(0), _width(1),_height(1) // default co
 {
 }
 
-int GameObject::getPosX(){	return _posX;}
-int GameObject::getPosY() { return _posY; }
-int GameObject::getWidth() { return _width; }
-int GameObject::getHeight() { return _height; }
+GLfloat GameObject::getPosX(){	return _posX;}
+GLfloat GameObject::getPosY() { return _posY; }
+GLfloat GameObject::getWidth() { return _width; }
+GLfloat GameObject::getHeight() { return _height; }
 
 
 void GameObject::Move(GLfloat x, GLfloat y) { _posX += x; _posY += y; }
@@ -27,7 +27,3 @@ void GameObject::MoveTo(GLfloat x, GLfloat y){ _posX = x; _posY = y; }
 void GameObject::Resize(GLfloat width, GLfloat height) { _width = width; _height = height; }
 void GameObject::ResizeWidth(GLfloat width) { _width = width; }
 void GameObject::ResizeHeight(GLfloat height) {_height = height; }
-
-
-
-
